requirements.txt
Django==5.2.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.8.0
django-environ==0.11.2
drf-spectacular==0.27.2
django-extensions==3.2.3
